Fix bug in DinPhil7
There was an assertion that I beleive is always violated:

a global `phil` is 0 initially, 7 threads are spawned which each
icrement `phil`, meaning the last thread to increment `phil` will leave
`phil` equal to 7.

The assertion checks if `phil` is 7, which is always violated.

Remove include for common.inc, I don't believe it is necessary.

Remove ECBMC function calls.